continue migrating to exposed
